+++ src/index.js
@@ -4,8 +4,9 @@ import { fetchMoviesMain } from './js/api/fetchMoviesMain';
 import './sass/main.scss';
 import { openModal } from './js/modal/openModal';
 import { createPaginationMain } from './js/pagination';
-import { renderGallery } from './js/help/renderGallery';
-fetchMoviesMain();
+import { renderGallery } from './js/gallery/renderGallery';
+import showMovies from './js/gallery/renderGallery';
+showMovies();
 
 const mainContent = document.querySelector('#main');
 // openModal => fetchMovieById => modalMarkup
